chore: rename `naming_service` to `name_service` for consistency
### CHANGES
- Rename `naming_service.py` to `name_service.py` in components.
- Update import path in `cli.py` to reflect renaming.
